requirements.txt
fastapi
uvicorn[standard]
langchain>=0.0.300
langgraph
transformers
torch
torchaudio
torchvision
diffusers
accelerate
bitsandbytes
coqpit
Pillow
python-multipart
google-search-results
TTS
gradio
pyngrok
soundfile
pydantic